requirements.txt
watchdog>=4.0
pystray>=0.19
Pillow>=10.0
pywin32>=306